+++ packages/components/organisms/f-header/CHANGELOG.md
@@ -4,6 +4,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+v9.2.0
+------------------------------
+*December 20, 2021*
+
+### Removed
+- 'ml' theme override for Logo.
+- width css definition from logo icon to make sure that icons are rendered correctly with one height.
+
+
 v9.1.0
 ------------------------------
 *December 14, 2021*